requirements.txt
blenderproc
numpy
objaverse
opencv-python
pillow
tqdm
trimesh